vol2birdR: Vertical Profiles of Biological Signals in Weather Radar Data

'R' implementation of the 'vol2bird' software for generating vertical profiles of birds and other biological signals in weather radar data. See Dokter et al. (2011) doi:10.1098/rsif.2010.0116 for a paper describing the methodology.

Details

To get started, see:

Dokter et al. (2016) doi:10.1098/rsif.2010.0116 : a paper describing the profiling algorithm.

'vol2bird' 'C' code documentation: an overview of the algorithm structure.

Lin T-Y et al. (2019) doi:10.1111/2041-210X.13280 : a paper describing the 'MistNet' model for rain segmentation.

See also

Author

Maintainer: Adriaan M. Dokter vol2birdr@cornell.edu (ORCID)

Authors:

Anders Henja anders.henja@gmail.com (vol2birdR package author and author of RAVE and HLHDF libraries)

Other contributors:

Alexander Tedeschi at744@cornell.edu (ORCID) [contributor]

Tsung-Yu Lin (contributor MistNet segmentation model) [contributor]

Subranshu Maji smaji@cs.umass.edu (contributor MistNet segmentation model) [contributor]

Daniel Sheldon sheldon@cs.umass.edu (contributor MistNet segmentation model) [contributor]

Bart Kranstauber b.kranstauber@uva.nl (ORCID) [contributor]

Jurriaan H. Spaaks (contributor to vol2bird library) [contributor]

Lourens Veen (contributor to vol2bird library) [contributor]

Iwan Holleman (contributor to vol2bird library) [contributor]

Hidde Leijnse (contributor to vol2bird library) [contributor]

John H. Merritt (author of RSL library) [contributor, copyright holder]

Bart Kelley (contributor and maintainer of RSL library) [contributor]

Mark Couture (author of iris2odim add-on to RAVE library) [contributor]

Daniel Falbel (contributor of original idea for building with torch support) [contributor]

Swedish Meteorological and Hydrological Institute, SMHI (copyright holder of HLHDF and RAVE libraries) [copyright holder]

GloBAM (https://globam.science) [funder]